Port of cudf::contiguous_split into spark_rapids_jni (#2799)
This is a setup PR for the implementation of `shuffle_split` and
`shuffle_assemble`. The split half starts with this version of
`cudf::contiguous_split` as a basis, so for ease of diffs with the main
PR I'm bringing this over separately. This is not called or in any way
used in `spark_rapids_jni` at the moment.

---------

Signed-off-by: Dave Baranec <dbaranec@nvidia.com>