GPU implementation of the Kudo shuffle format. (#2800)
This PR provides the GPU implementation of the split and assemble halves
of the Kud0 shuffle format. It is dependent on
https://github.com/NVIDIA/spark-rapids-jni/pull/2799. That PR exists
solely to help reduce the amount of code that needs to get reviewed for
this PR. `shuffle_split.cu` should be a little easier to review once it
gets merged.

`shuffle_split` implements the sender side half of the Kud0 format.
Splitting an input table up into the final output binary and associated
metadata. `shuffle_split.cu` is where it is implemented, and is based on
`cudf::contiguous_split`

`shuffle_assemble` implements the receiver side of the process -
reassembling one or more blobs produced by one or more `shuffle_split`
calls.

This is feature complete and in a reviewable state, but there is a known
issue with `struct<list>` and `struct<string>` columns that I have to
fix. I don't expect it to alter the code much. I also need to add more
tests. But in the interest of getting reviews going, I'm putting this up
in draft form.

---------

Signed-off-by: Dave Baranec <dbaranec@nvidia.com>