Fix `GpuTimeZoneDB#isSupportedTimeZone` (#3169)
Closes #3170 
Fix CI break, previously isSupportedTimeZone was set to what it should
be after https://github.com/NVIDIA/spark-rapids/pull/12502 is merge
(that the timezone is a valid one), but since we want to support the
same behavior, need to set it in the same way that it was before.

Currently, Spark-Rapids wants all DST timezones to fall back to the CPU,
this behavior is currently broken, it is not falling back. This PR will
fix that behavior, and it will fall back as intended.

To reproduce:
TZ="America/New_York" ./integration_tests/run_pyspark_from_build.sh -k
date_time_test.py -m tz_sensitive_test

---------

Signed-off-by: Warrick He <warrickh@nvidia.com>