Fix compile warnings when comparing unsigned char to zero (#3242)
This fixes the (potential) warnings when comparing unsigned char to
zero. Most of the times, the warnings do not show up because `char` is a
signed type on `x86` platform but it can be unsigned on `arm64` platform
(`char` can be signed or unsigned, it is not specified by C++ standard).

Fixes https://github.com/NVIDIA/spark-rapids-jni/issues/3227.

---------

Signed-off-by: Nghia Truong <nghiat@nvidia.com>