Java APIs for GPU kudo (#3100)
Please note that this updates the CPU format to match the GPU format for
kudo in places where the GPU appears to be more self consistent or the
size of the data was smaller.

The main changes are padding and buffer order. Padding only happens at
the end of each section, instead of in between each buffer. The order of
the buffers for a struct has changed so that the child buffers come
after the buffers for the struct. This makes it consistent with lists,
which also had them come after. I also updated the code so that if a
validity buffer was skipped because there were no rows, then the flag to
say if the validity buffer is there or not indicates that it is not
there. Before the CPU code would set it to say that the buffer was there
in some cases, but know that it was not there because the row could was
0. This just makes it more consistent and should be matching what the
GPU is doing.

---------

Signed-off-by: Dave Baranec <dbaranec@nvidia.com>
Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>
Co-authored-by: Dave Baranec <dbaranec@nvidia.com>