Change the priority to be first seen is highest (#3290)
This does not really have a big impact yet. I ran NDS and saw no
improvement or loss, even if I reduce the amount of memory available to
the processes by a lot. The improvements will come when I get spill to
also have the same priorities, and then we start to run into situations
where there is a lot of spilling.

---------

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>