Add design doc for memory tracking in the plugin (#2628)
This PR adds a new doc explaining the design of the memory tracking
system in the plugin. The new doc is based on the existing doc written
by @revans2, which is no longer publicly accessible, and fixed to be
up-to-date. The main purpose of this PR is to have a pointer to the
design doc in the code as
[here](https://github.com/NVIDIA/spark-rapids-jni/pull/2628/files#diff-936903e8604f063e1364ecc9ad83dcd4d5552d6c26d93073b469f99ca562d7a3R84),
so that people who are interested in this area can easily access to the
design doc.

---------

Signed-off-by: Jihoon Son <ghoonson@gmail.com>