Fix bug in cast string to timestamp: just time string without leading 'T' is also valid (#3333)
## bug
"01:2:3" and "1:2:3" are also valid just time strings besides "T01:2:3"
and "T1:2:3"
Before this PR, the code only handled 'T' leading strings.

## fix
Check if 'T' is leadning, and if the 2nd or the 3rd character is ':', if
it is, then try parsing time.
Note: sing '+' and '-' should not be leading just time, e.g. the
following is invalid:
"+T01:2:3"
"-1:2:3"

Signed-off-by: Chong Gao <res_life@163.com>

---------

Signed-off-by: Chong Gao <res_life@163.com>
Co-authored-by: Chong Gao <res_life@163.com>