Apply change to HostTableJni to remove confusion in cuDF builds (#3405)
There has been some confusion in cuDF ci (see
https://github.com/rapidsai/cudf/pull/19045 for an example) on why their
spark-rapids-jni build is failing. The changes stem from this PR
https://github.com/rapidsai/cudf/pull/18983, which refactored
headers/includes causing a .cpp file in spark-rapids-jni to fail.

Additionally, the build currently from cuDF is testing cuDF 25.08 vs
spark-rapids-jni 25.06 (since we are "in between" versions right now).
This is something we'll address in the future.

This PR is an attempt to make the spark-rapids-jni build "green" even
though it's not blocking cuDF, to hopefully reduce confusion overall.

Signed-off-by: Alessandro Bellina <abellina@nvidia.com>