Fix overflow in decimal scaling operation (#3452)
When doing decimal scaling, the scale factor (256bit) is incorrectly
cast into `int` value, which cannot hold the 256 bit value. This fixes
it, keeping the 256bit value without casting for further operation. A
new Java test is added to verify this.

This also fixes a few memory leak issues in Java tests, and renames
variables into a more accurate meaning.

Closes https://github.com/NVIDIA/spark-rapids-jni/issues/3451.

---------

Signed-off-by: Nghia Truong <nghiat@nvidia.com>