Revert "scl default" changes made previously given CI failures (#3474)
Closes https://github.com/NVIDIA/spark-rapids-jni/issues/3299

The issue here looks to be that we are enabling a devtoolset from within
a shell that already has devtoolset enabled. TableTest, in particularly,
will shell out (invoking bash, see
https://github.com/apache/hadoop/blob/trunk/hadoop-common-project/hadoop-common/src/main/java/org/apache/hadoop/util/Shell.java#L852),
and this is the stack where we hang.

It looks that enabling toolsets should be done once, and this nested
case may be error prone. I have reverted the change locally and cannot
reproduce the hang anymore. I believe this to be the root cause of our
hangs in CI.

---------

Signed-off-by: Alessandro Bellina <abellina@nvidia.com>
Signed-off-by: Nghia Truong <nghiat@nvidia.com>
Co-authored-by: Nghia Truong <nghiat@nvidia.com>