Expose the row index of one failed row when casting floats to decimals. (#3479)
contributes to https://github.com/NVIDIA/spark-rapids/issues/11550

This PR exposes the row index of one failed row when casting floats to
decimals to replace the `has_failure`. This index will be used to get
the row value which will be used to build the excepion message, to align
with the behavior of Spark.

Spark will return nulls for NaN, Infinity and -Infinity even the ANSI
mode is on, so this PR also changes the code to follow the logic.

---------

Signed-off-by: Firestarman <firestarmanllc@gmail.com>
Co-authored-by: Nghia Truong <7416935+ttnghia@users.noreply.github.com>