Use the actual digits count for the overflow check when casting strings to decimals (#3521)
This is a fix to the issue
https://github.com/NVIDIA/spark-rapids/issues/10908.

`significant_digits_before_decimal` is the digits count of the stored
value, not the actual value. When the scale is negative (in Spark), the
stored digits count is not always equal to that of the actual digits.
For example, the digits `123456` with DecimalType(4, -2) has `1235` as
the stored digits, but the acutal digits is `123500`.

The overflow check should use the actual digits count instead, which can
represented as `decimal_location - leading_zeros`.

More tests for this change are added in by the Plugin PR at
https://github.com/NVIDIA/spark-rapids/pull/13089.

---------

Signed-off-by: Firestarman <firestarmanllc@gmail.com>