Add in support for 64-bit aggregation utils to help with SUM overflow detection (#3531)
This code is based on the 128 bit code in CUDF that does similar
processing.


*
https://github.com/rapidsai/cudf/blob/branch-25.08/java/src/main/native/src/aggregation128_utils.cu
*
https://github.com/rapidsai/cudf/blob/branch-25.08/java/src/main/native/src/aggregation128_utils.hpp
*
https://github.com/rapidsai/cudf/blob/branch-25.08/java/src/main/native/src/Aggregation128UtilsJni.cpp
*
https://github.com/rapidsai/cudf/blob/branch-25.08/java/src/main/java/ai/rapids/cudf/Aggregation128Utils.java
*
https://github.com/rapidsai/cudf/blob/branch-25.08/java/src/test/java/ai/rapids/cudf/Aggregation128UtilsTest.java

It is probably a good idea to start off by comparing this code to that.

I am not sure how permanent this code will be. I am hopeful that CUDF
can have a version of https://github.com/rapidsai/cudf/issues/19243
working by the next release.

---------

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>
Co-authored-by: Nghia Truong <7416935+ttnghia@users.noreply.github.com>