[Refine] demonstrate using NVTXW to export internal CUPTI events to nsys-rep (#3130)
Refinement to https://github.com/NVIDIA/spark-rapids-jni/pull/2450.

originally description:
> This branch adds support for exporting internally collected CUPTI
events using NVTXW. NVTXW is an API that allows applications to record
their own profiling events and then insert them into an Nsight Systems
report (nsys-rep file) in a post-processing step.
The README-nvtxw.txt file outlines the steps to use this feature. The
"libNvtxwBackend.so" library mentioned in the readme is included in the
Nsight Systems release. The latest release is available here:
https://developer.nvidia.com/nsight-systems/get-started. Nsight Systems
2024.6.1 should be posted in the next few days, which is the version I
used for testing. In the mean time, 2024.5.1 may also work, or I can
provide a link to an internal build server with pre-release build of
2024.6.1, if that would be useful.
Marking this feature as WIP and draft. It works as is but is intended to
be a demonstration of how to use NVTXW. Feel free to refine, refactor
and improve this code as appropriate.

This pr addressed some comments, use NVTX 3.2, and fixed a dependency
issue based on 2450.

<!--

Thank you for contributing to RAPIDS Accelerator for Apache Spark!

Here are some guidelines to help the review process go smoothly.

1. Please write a description in this text box of the changes that are
being
   made.

2. Please ensure that you have written units tests for the changes
made/features
   added.

3. If you are closing an issue please use one of the automatic closing
words as
noted here:
https://help.github.com/articles/closing-issues-using-keywords/

4. If your pull request is not ready for review but you want to make use
of the
continuous integration testing facilities please create a draft pull
rqeuest
   or prefix the pull request summary with `[WIP]`.

5. If your pull request is ready to be reviewed without requiring
additional
   work on top of it then remove any `[WIP]` prefix in the summary and
   restore it from draft status if necessary.

6. Once all work has been done and review has taken place please do not
add
features or make changes out of the scope of those requested by the
reviewer
(doing this just add delays as already reviewed code ends up having to
be
re-reviewed/it is hard to tell what is new etc!). Further, please avoid
rebasing your branch during the review process, as this causes the
context
of any comments made by reviewers to be lost. If conflicts occur during
review then they should be resolved by merging into the branch used for
   making the pull request.

Many thanks in advance for your cooperation!

-->

---------

Signed-off-by: Haoyang Li <haoyangl@nvidia.com>
Co-authored-by: Tod Courtney <tcourtney@nvidia.com>
Co-authored-by: Nghia Truong <7416935+ttnghia@users.noreply.github.com>