Fix task priority issue with test order (#3698)
When we switched from using the task id as priority to using FIFO order
for priority two tests were missed in the update. They still passed, but
this was mostly because previous tests ran first, set the correct
priority for the tasks and did not clear the priority when they were
done. If `testBasicBUFNCpu` or `testBasicBUFN` were run on their own the
test would fail.

This fixes a lot of small issues with the tests.

1. It makes sure that when a task thread is created the priority for the
task follows the order initialize was called.
2. It makes sure that when a task thread finishes the task thread is
cleaned up properly
3. It assigns a new task id for each test so that we can avoid
bleed-through if a similar bug shows up in the future.
4. It updates all test, not just those with task threads, to clean up
properly
5. It fixes `testBasicBUFNCpu` and `testBasicBUFN` so the comments are
correct and the test initializes things in the proper order.

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>