Package nvmljni into a separate .so which can be optionally loaded when available (#3899)
Fixes https://github.com/NVIDIA/spark-rapids-jni/issues/3891

Package nvmljni into a separate .so which can be optionally loaded when
available, otherwise gracefully fallback to no-op behavior. This PR
depends on https://github.com/rapidsai/cudf/pull/20414

Tested this by building and running `NVMLTest.java` and then also
rerunning with the `libnvidia-ml.so` effectively hidden on the system.
Before, this would fail on loading the regular native deps (because
libcudfjni.so itself depended on libnvidia-ml.so), but now it passes in
both cases.

---------

Signed-off-by: Zach Puller <zpuller@nvidia.com>