Decouple logger object from spark_resource_adaptor (#3931)
Contributes to https://github.com/NVIDIA/spark-rapids-jni/issues/3905

This centralizes log_enabled and log_level checks inside static macros
we can use in the adaptor, and adjacent classes.

Having the logging managed this way allows us to not execute std::format
lazily consistently, although it was almost perfect before. That said,
it makes the logging calls more fool proof, and allows us to have
specialized macros that can log containers for example.

This also allows us to move task state specific logging to the task
state classes, without having to pass the logger around, or the adaptor
(we can use the static logger).

LOG_INFO and LOG_DEBUG are not currently used, but they are useful for
debugging.

---------

Signed-off-by: Alessandro Bellina <abellina@nvidia.com>