Support nanosecond precision in timestamp conversion (#4095)
Closes #4089. 

This adds support for nanosecond precision for timestamp conversion for
`fromUtcTimestampToTimestamp` and `fromTimestampToUtcTimestamp`. The use
case is parsing timestamp strings -> converting between timezones ->
storing as strings, preserving nanosecond precision.

The change is limited to GpuTimeZoneDB since it is hard to replicate
outside the JNI, whereas e.g. CastStrings is easier to replicate using
plain cuDF parsing. But in the future, supporting nanos in CastStrings
`toTimestamp` would be desirable.

---------

Signed-off-by: Rishi Chandra <rishic@nvidia.com>