Move hash-related code into `src/hash`
Fixes #4104.

This is a code hygiene task that came up as part of
https://github.com/NVIDIA/spark-rapids-jni/pull/4091.

The `src/main/cpp` directory is beginning to get crowded with code that
handles a multitude of concerns.  It would be good to get this a little
more organized.

As a start, this PR moves the hash-related functionality
to the `hash/` directory (introduced in #4091).  This would make it
easier to browse and locate related code.

Signed-off-by: MithunR <mithunr@nvidia.com>